skills/gripper_commands: Mock RelMessage by skill
Before the MOVEREL command was directly linked to a MoveRel message
of the arduino plugin. As the arduino plugin just ignores any movement
outside the limits, instead of clipping it, this leads to problems
during MPS handling.

This commit now solves the problem by clipping the given relative values
into the limits, allowing movements as far as possible.