skiller_motor_state: Use mutex to get rid of race condition
When the thread waits for the timeout it can be stopped by the data
changed callback routine. However, just before the wait method is
called, the callback could run without stopping the timeout, leading to
an delay of around the timeout time.

By employing mutexes, this potential future issue could be already
resolved now.