tag_vision: fix transform publishers
tag_vision also makes the very common mistake of using a single
TransformPublisher to publish multiple transforms. This is fine if the
transforms aren't bbsync'ed. The transforms themselves get published
fine, but there is only one blackboard interface that is continuously
overwritten with the transform data. This makes it impossible to bbsync
them.