cx: do not sync the robot's position
Syncing the robot's position may cause problems if we remove and
reinsert the robot, as its position would be synced back from the other
robots, often resulting in two 'at' facts for two different positions.
As we do not need the other robot's position, simply do not sync it.

Also re-initialize the fact such that the robot is back in the insertion
zone.