cx: use goal resources instead of SPAWNING-MASTER for WP-SPAWN
Use a run-endless goal with retract-on-REJECTED and do not reformulate
if the mutex is already locked. This way, all bots try to create a
WP-SPAWN-MAINTAIN goal, but only one of them will succeed. The other
bots will retract the goal and won't retry until the mutex has been
opened again.