cx: only award points when the step was not performed already
The tracking system only awards points, if there is yet an order to be
fulfilled that needs that step AND this step was not performed yet. This
means, if we lose a product, before we deliver it, the already scored
points are not scored again if we retry to build it. Exceptions to this
are:
1. If the quantity requested of the product is > 1, then we could gain
the points for the intermediate steps of the second delivery of said product.
2. If some other order requires the same production step and has the
same previous production steps. Then points are awareded for the
intermediate steps of that other product, if that did not happen
already.

For now, we ignore case 2 in our model.