cx: prevent potential rs deadlock
When a C2 or C3 product is at the output of a ring station and needs to
get a ring mounted from the same machine, we should keep the input of
said machine free. Otherwise another CX product might mount a ring on
the station while we clear the output and thus 'steal' the additional
bases required to mount the ring on the C2 or C3 that was on the output.