arduino: Make gripper status checking better
The decision whether the open/close command should be followed is not an
easy one.
If for example, the gripper is open and the gripper is opened again, the
belt slips, which lead to mechanical erosion and makes future gripper
less reliable.

This commit introduces the usage of an assumed state of the gripper, as
well as the not so reliable end stop state.

The following cases are possible:

assumed_state end_stop   reactions

open          open       no open  @ open cmd
                         close    @ close cmd
closed        open       open     @ open cmd
                         no close @ close cmd
open          closed     open     @ open cmd
                         close    @ close cmd
closed        closed     open     @ open cmd
                         no close @ close cmd