protoboard-rcll: factor out domain definition
protoboard-rcll is effectively a new plugin, but it's based on the
protoboard plugin. Only this time, it contains ONLY the domain-specific
template specializations & instantiations. All of the generic (i.e.
domain-agnostic) functionality has been moved to a plugin template in
the core repo.

This compiles, but is COMPLETELY UNTESTED yet.