gazsim-gripper: publish gripper_dyn_* transforms
since the gripper_command skill now depends on the time-stamp of the
"gripper" frame to determine if the result is ready, publish the needed
transforms. For the tfs are just published with no new value just to
cause the time-stamp update (to be newer than the time-stamp of the
ArduinoInterface).

Later on (if we decided to finally actually simulate the gripper movement)
further adaptation to expected positions will be published.
Some vars are not currently used (like curr_x_ and moving). They will be
used later on when simulating the gripper movement.