cfg gazsim overrides: overrides for arduino max_* and min_* allowed coordinates
The values are used by conveyor_algin skill to determine the tolerance
of the gripper movment from the current conveyor_pose. This is heavily
used to control the state transtion between DRIVing and LOOKing and their
corresponding STOP_* states.

The desired behavior is that after reaching a pose where the gripper
can do the rest of the movment the alignment succeeds and lets the
gripper movement cover the rest of the distance.

since we do not yet simulate the gripper movement, the chosen values
should make sure to not stop the alignment pre-maturely yet still allow
the robot to get close enough to the machine.

The current values work ruffly all the time. Yet revisiting them is
needed (especially when adapting the gripper movement)