gazsim-gripper: The gripper_y_dyn is at 0,0,0 of its parent (gripper_y_origin)
This was causing the tolerance_ok function() in converyor_align
to never success while checking the gripper's min and max y tolerance.
(One of them always failed ,either min or max).
Both condtions are calculated from the cgf value max_y, which our y_dyn
is also relative to. Wondering how is this working in the realworld