mps-laser-gen: refactor to avoid code duplication
The MPS objects were computed twice if both the box filter and the
laser-gen config options were enabled, and most of the code to do that
was duplicated. Refactor such that we only generate an MPS object once
and use it for both options. This also gets rid of the duplicated code.