gologpp: add a multi-agent visitall domain with a central agent
This does not really visit all locations, but just lets the first robot
move to the C-BS and the second to the C-CS1. It runs a central golog
agent that uses the remote skiller executor to execute the actions on
the robots. Each robot only runs the skiller, no agent.