cx: recursively expand production subgoals on selected parent
We used to expand the whole tree in 2 steps. Once on the selection of
the production-maintain, generating all compound subgoals. Then we
formulate the simple goals (at leaves) when their compound parents are
generated.

The inconsistency in tree expansion semantics is kind of confusing
specially when it involves algorithms encoded as rules to switch
the node parents of a simple goal.

A simpler more consistent approach is to expand the branches of the tree
when the parent is selected. Since our production tree consists mostly
of run-one compound goals. This will select the highest priority branch
(ie, depth first manner) till we find a goal we commit to