cx: only run action if parent of simple goal is dispateched
Since we dispatch bottom up currently, we do not want the simple goal to
be evaluated before its parent is dispatched. This is a work around for
the race condition that a simple goal is evaluated too fast.

TODO: a more general approach where in a tree.. A goal can only be
evaluated if its parent has been dispatched.