tests: use mktemp to create the temporary config overwrite
Writing to `host.yaml` does not work because the file is not read when
using the gazsim startup script. Instead, just write the config to a
temporary file in `conf.d` and delete the file when stopping the test.
This also ensures that we do not accidentally leave the simtest enabled,
or overwrite any other config values.