cx: sync the keep-mps-side-free fact
This fact blocks the formulation of MOUNT-xxxx-RING goals that target a
RS input side, where the output is occupied by a product that needs the
same input next.
Only the robot that executes a PROCESS-MPS goal can assert this fact as
it is done as a pre-evaluation step. Therefore, this fact has to be
synced in order to properly prevent possible deadlocks at ring stations.