gazsim-scripts: wait for the robots to send their beacon
Sometimes, the robots need longer than 10 seconds to start. To make sure
all robots are there, do not sleep for a fixed amount of time, but
instead wait for them to send the beacon.

We still need a (smaller) sleep to make the refbox SETUP phase has
completed before switching to production.