clips-mip-scheduler: constraints 2,3 and 5,6
Equations (2) & (3),ensure the material balance property at the source and
sink events of a commodity.

(2) models that the in-flow of a resource r to the sink event i, must satisfy
the event’s requirement of that resource (i.e., the total number of selected
incoming edges of some resource type X r,i,j , must equal
the resource requirement E ri of event i ).

(3) models the same property for the flow, of a resource r, out of the source
event j (i.e., the total number of outgoing edges,selected in the solution, of
some resource type X r,i,j must equal the supply E rj of event j).

(5) & (6) ensure the material balance property at the event nodes that belong
to a plan of a goal. They express that commodities are allowed to flow
inkout of an event, only if the plan that contains the event is selected.
In that case, the number of commodities flowing inkout needs to satisfy the
requirement supply of that event from the commodity type.