tests: do not export ROS logs to the local log directory
ROS creates an absolute symlink, which causes buildkite failing to
upload the build artifacts. The ROS logs are not interesting anyway, as
we do not actually run any ROS software, but only export data to it, at
least in the tests. Therefore, do not export the ROS logs to the output
directory.