cx: add default termination field to testcase
This gives a convenient way to have tests that passively check whether
some condition happens or not without interfering with the runtime of
the tests. Custom termination means that it is up to the test writer to
specify the conditions under which a test may fail or succeed. The other
termination types simply terminate once all custom terminations are
done. An example use-case is the FLAWLESS-MPS test which may be relevant
both for short tests that only run partial games and full game runs.