cx: scheduling specific deftemplates to track a schedules life cycle
Fact "schedule" keeps track of the mode (similar to goal modes) and the
goal for which the subtree will be scheduled.

Fact "schedule-event" tracks the events used by a schedule by parsing
all plans, sub-goals resource usage of a goal-tree into corresponding
at START and END events to be passed later for the scheduler.

Fact "schedule-resource" tracks resource used by a schedule, wither
they are consumable or producible within the schedule (by a plan),
the setup states possible for the resource and the units available.

Fact "schedule-requirement" tracks resource required by an event of a
schedule. The units consumed of produced by the event and the required
setup.