cx: introduce a 'schedule' lifecycle On EXPANDED SCHEDULE goal
Schedule Lifecyle:
Schedules are direcly related to a goal-tree with sub-type SCHEDULE-SUBGOALS.
Upon expansion of the top most goal with SCHEDULE-SUBGOALS subtype (i.e., the
schedule root goal), a 'Schedule' fact is created tracking a single scheduling
attempt (i.e, a single call to the scheduler). A Schedule has lifecyle
similar to goal-lifecycle, tracked by its
modes.

FORMULATION [a schedule is created for the schedule root goal]
 - all sub-goals in the 'SCHEDULE-SUBGOALS' subtree are added to the
   schedule (goals)
 - schedule-events and resource requirment are extracted from the tree
   and the plan resource requirments
 - the scheduler is called to build the datasets

SELECTION [the scheduler is called to trigger the optimization]
    - Process the results of the optimization into corresponding schedule-event

EXPANSION [a schedule has been found]
    - allows reasoning if this schedule should be COMMITTED for executing the
       goal tree
COMMITMENT [create setup goals needed for execution, at the right branch of
             the scheduling goal-tree]

DISPATCH [A schedule is ready to be executed, ground the starting-time]

Upon dispatching of a schedule, the scheduling goal-subtree continue its
lifecycle, according to the dispatched schedule]