cx: late binding for FVAR_WP, rule for grounding "wp-spawend-for"
This is a temporary solution to the late binding issue. The problem here
is that the MOUNT-CAP and DELIVER plans need to use a WP grounded by a
predicate (Wp-spawned-for WP ?R) where the robot is not known when the
plan is constructed.. The robot selection (which would happen by the
scheduler) would later determine the binding of the WP.

More generally, I need a consistent formulation to solve the late binding
problem by allowing the symbolic reasoning on a set of temporary created
free-vars. The vars are created with specification on how to ground them
(eq, CEs that would eventually match, The grounding of the vars could happen
later when a set of facts match which should in turn trigger unification).