cx: adapt to new goal-tree, propagating goal start from plan start
It used to be that a parent start time will be the least of its
children. That preserved chronological commitment and dispatching
(top-down).

With the new goal-tree structure, the dependency is rather logical between the
sub-goals, such that subgoals dispatch and complete before parent is
committed. The start time of each goal is the start time of its plan