cx: least commitment to params via meta binding facts
Sometimes binding to a certain paramters needs to be delayed to last
moment possible. This saves some decesions from being made by the
scheduler in the cases where the specif allocation is not required.
We provide a way to achive this by binding creating a
(wm-fact (key meta binding-id ?id wm-fact-key $?predicate))

where predicate is the key of the wm-fact that needs to be matched later
on