cx: meta binding, refactoring of meta binding fact and functions
A meta binding fact is created when the binding to a wm-fact needs to be
delayed till execution. mate Binding facts has the formate of
(wm-fact (key meta binding args? id Xgen123 policy BIND-UNIQUE|BIND-ANY)
         (values $?wm-fact-key))

Where the wm-fact-key is the partaily bound predicate to be matched
later on.

Policy defines if this fact is allowed to be bound to any matching
wm-fact or does is it required to bind it to a fact which was not
previously bound. A use case for that is binding UNIQUE binding is
binding to consumable resource that should only be used once.
A use case for ANY is those state predicates that might occur several
times or be used multiple times (eg, rs-in args? sum ?sum summand ?)