cx: exec remote-action working scheme
For plan-actions that are meant to be executed by a different bot (ie,
with an (r ?robot) parameter different than ?self and/or identity).
Remote execution is achieved by syncing of wm fact describing the remote action
(wm-fact (key exec remote-action args? id ? goal-id ? plan-id ?)),

On the reasoner side:
 - plan-action PENDING & EXECUTABLE --> create the wm-fact
 - wm-fact is updated --> update the plan-action
 - state is EXECUTION-SUCCEEDED/FAILED --> remove wm-fact

On the executor side:
 - wm-fact with no corresponding plan-action --> Create plan-action
 - plan-action is updated --> update wm-fact