cx: deduce resources and thier at-start/end states from plan
A PDDL object type resource is declared objects that should be
considered as such.

After a plan is created and the domain-preconditions are grounded.
Deduce the resources used by a plan and their state predicates at the start
and at the end of the plan.

start-state-set: Is the set of all postitve preconditions that were not
satisfied by a positive effect of an earier action in that plan

end-state-set: Is the set of predicates achieved by a positive effect of
an action, where that effect was negated  by another action later in the
plan.