clips-mip-scheduler: CONTINUOUS instead INTEGER time vars
A very strange problem happens here. The comparison when constructing the
schedule-resource such that the events are arranged by their start time
from smaller to bigger...
The float comparison fails sometimes with no known reason comparing an
event starting at 0 with an event starting at a later time.
This is such a strange issue which all the other sides had been rolled
out. The problem is clearly that the float comparison does not always
give the expected result. I suppose this might has something to do with
float point representation in different  langs