clips-mip-scheduler: General Event-Selector semantics (rp goal,plan sem.)
This is the first step to apply a more general event selection semantics
that works for all events instead of the old goal plan relation.

We are trying to arrive to away to generally represent that an event
could be selected and when would it be selected.
The old semantics relays on the facts that
A goal is always selected
A resource is always selected
A plan should be selected for a goal. (P in G)

The new semantics partially introduces the concept of Event Selectors
and Selector Groups.
-The idea is that all events have a selector (as opposed to only some from
 old model)
- An event selector could be TRUE to begin with in which case its always
  selected
- The Events selectors could belong to groups, which could also be in
  turn selected
- A group has a selection policy to the events it hosts, as in
  SELECT_ALL or SELECT_ONE. This will be used to control the grouping
  constraints  (previously the TotalGoalPlan)

This is a partial implementation, only moving the semantic from plans
and goals simply  mapping it to what in means for selectors
- all events have selectors
- on add-goal the selector is set to TURE
- on add-event the selector is set to FALSE
- on add-goal-plan a grouping is added for the selector