gazsim-scripts: quit early and stop sim if robots do not start
This was supposed to happen already, but `rcll-refbox-instruct` does not
allow a space between `-W` and its argument (i.e., `-W60` works, `-W 60`
does not). Also remove another sleep that is unnecessary.

If the robots fail to show up, actually stop the simulation and quit.
Without this change, we started the game anyway, even if the robots did
not connect.