cx: setup links only if requirments identical (not working as expected)
The idea is to only allow the setup links to be created between events
releaseing and consuming a resource in the same states.
Several problems showed up..First of all the code is not working as
expected. But moveover there is a problem with the concept itself since
the resource-states also includes undesired predicates like wp-on-shelf.
Moreover, the grounding of the predicates with the specific resource
is crusial. The issue is that we care not about the specifiec resource
but rather a type