cx: resource re-definition  [stage 1]
* Alter "resource" fact to hold resource-type related info, including
  setup-predicates and state-predicates

* Alter "schedule-resource" to hold the mapping of
  (entities Vs. ids Vs. types), besides the old event-seq

* schedule-requirement is extended with two multislots, resource-states
  and resource-setup. Each holds a collection of statements about the
   requirement of the event

* schedule-requirement is deduces from the plan-processing information
  (i.e., plan-resources), basically splitting the resource statements
  into resource-setup and resource-states.

This commit is back to the old functionality of only using the
requirements to deduce setup-edges. Following we will extend to only
create setup-edges for resources-consumption/production with the same state.

Along the way:
resource-info and setup calculations is refactored and moved to another
file (resource-sched.clp). A usful statment processing functions added