cx: match mps-handling facts with the right goal
The mps-handling facts are asserted in the goals that transport the
associated wp to the target mps. Since we might bring a product to the
DS and then pick it up again in favor of delivering a different product
first. In those cases multiple mps-handling facts are present for the
same machine. In order to match the correct mps-handling facts in this
case it is necessary to match the remaining goal params to the
mps-handling params.