cx: schedule, use resource-states for in setup-edge gen. [and analysis]
Only generate setup-edges for a resouce if the consumer-resource-state
is a subset of the producer-resource-state.

This does not work because the "resource-state" describes only the
effects of a plan actions, rather the state of the resource.

For example,
A CLEAR_CAP plan will remove the 'wp-at' and change the mps-state from
READY-AT-OUTPUT to IDEL, yet it has no mention of any of the states the
resource was in before the plan had started (like can-perform,
cs-buffered, mps-state)
Indeed, what we need is the full state of the resource at the point when
that plan finishes.

For a strict sequence of plans, that would be similar to what deduce
plan effects, that is, the resource-state is all the positive effects of
previous plans that were not negated by the current effects of the plan
in addition to the new positive effects of the plan.

The question that rises is what is there is more that one possible
sequence of plans. Some thing like
Plan1 --effects[ A B ]
Plan2 --effects[A c]

CurrentPlan preconds [ A ]

Should we include both B and C in the resource-states of CurrentPlan
effects?

For now I think that yes as the resource-state needs to describe the
'possible' states a resource could be at this stage, to determine wither
which next plans can we perform..
For instants after the CurrentPlan it should be possible to any of the
plans 3 and 4:

CurrentPlan effect [ ~A ]

Plan3 --precond [~A , B]
Plan4 --precond [~A , C]

Meaning, there should be a setup edge from CurrenPlan to Plan3 and
Plan4. The choice of which one to pick needs to be determined by the
scheduler.