clips-mip-scheduler: less_then_eql +1
When the durations of events is zero, it could be that loops happen
because the precedence constraints no longer insure the flow in one
direction only  (E1 + e1dur == E2 + e2dur).

To remedy we could use a stricter "less_than" relation, yet its not
allowed by gurobi. To get around this we ensure there is at least one
second in between events.