cx: scheduler, Introduce edge resource-states and  edge flow groups
The previous assumption about accumulating a resource-state from one
event to another was dump and had several issues. The least was that the
resource-state no longer means that 'expected state' at that point, but
rather the possible states which lead to this event.

Here, we introduce a much more concrete concept edge resource-states.
After finishing an plan-event, a resource will have the state which it
had before the plan modified by the application of a plan effect.
That is, knowing the state of a resource before a plan one can certainly
make a statement about its state after.

*Each resource-edge, out going from a source event, has the initial state
of the resource.
*A resource-edge incoming to an intermediate event with a certain state,will
have a corresponding out-going edge with that state after accumulating
the effects.
*An edge is extended to an event iff the preconditions of the event is
satisfied by the edge resource-state.
*All resource edges incoming to an event, which cause the same edges
resource-state to come out will have a flow equivalent

----------                   -------
|source1 |  ==A,B,C========> |      | ===B,C,X,Y==>
----------                   |Inter-|
                             |media-|
----------                   |te    |
|source2 |  ==A,D,E========> |Event | ===D,E,X,Z ==>
----------                   --------
                          Precod (A)
                          Effect (~A,X,Z)