cfg: revert team-name in agent.yaml. define as cx/rcll config
The team-name is more related to the agent implementation than the
agent itself. This also avoids the redundancy in having to specify the
team for every host. Its also easier to see in relation to the refbox
configs declared in the cx.yaml

This commit reverts part of the changes introduced by 5fd913a63