cx: prevent formulating mount goals on products with caps
The main reason why this did not cause troubles earlier is, because the
product remained at the cap station output until delivery, hence the
machine was never buffered with a new cap.
Now we bring products to the delivery station nearly, hence it can happen
that a finished product is at the DS while the respective cap station is
fed with a buffered cap again. This allows to mount the cap again if we
formulate PRODUCE-CX goals regardless of the current cap color on the
target workpiece.