scripts: fix gazsim.bash tmux invocation
Can't start tmux detached because the refbox shell needs to know the
terminal size.

When spawning our windows inside an existing tmux session, we need to
rename it.

Also, on '-x kill' we don't want to kill the entire tmux session, just
the individual processes like in the non-tmux case.