cx: replace clear-cs expired with clear-cs when needed
Part of a new two stage approach to handling finished products, instead
of leaving a product in the CS until it can be further processed or is
discarded when it is expired, take the product out of the CS, when a
product is placed on the input side. Then handle the product according
to downstream rules.