gazsim-scripts: enable bbssynced robots on central agent startup
The central agent dispatches commands via the remote blackboard to the
acting robots. This requires to open the connection through
configuration of reading and writing interfaces, but only for the
participating robots. Hence the config depends on the number of robots
that are specified on startup.

Simply write a config file on startup that specifies the active robots
and that can be loaded along with the central agent config.