rcll-central: add comments on bs instruction workflow
The workflow for interacting with a base station expects that a
corresponding transport goal is started and the robot is ready to pick
up the product. Only then the BS is instructed. This lazy instruction
policy ensures that if multiple robots intend to use the base station at
the same time, the one that arrives faster gets precedence.