rcll-central: increase salience when selecting a root goal
This is just a workaround: The agent has problems transitioning from
exploring machines to producing orders. One problem was that the
agent rather selected subgoals of a run-parallel goal instead of
selecting a fresh root goal.

I think there should be a cleaner workflow all together in the long run.