goal-class: update goal-class assertion to decrease complexity
Enforce a more strict separation between preconditions that are necessary
for formulating a precondition formula for goal formulation (i.e. accessing
facts that hold information for a particular order, WP, etc.) and those
that are required for actual goal formulation (i.e. transient facts that
represent a world-state and need to be fulfilled to actually execute the
goal). The former are captured in the form of CLIPS LHS preconditions
before the goal-class for the corresponding goal is a asserted and serve
configuration purposes. The latter are enforced in the form of the
precondition formula that is continously evaluated, once asserted and
grounded, against the current state of the world-model.

The existing goal-class create and assert goals are updated accordingly.