rcll-central: do not delete rs-ring-spec for RING_NONE
This fixes an issue, where order meta information is not being created
when a C0 challenge is being run.

Then a race starts after the order info is received.
The next machine info deletes the rs-ring-specs for RING_NONE, while the
initialization of order meta information is a low-priority rule that
relies on rs-ring-specs for RING_NONE being present.