xenonite: use correct timepoints for action preconds and effects
Do not always use `at start` for the precondition and `at end` for the
effect, but instead properly model the actions. As an example, when
picking up a container, the robot must be at the location for the whole
duration of the action, not only at the start.

For the move actions, remove the start location at the beginning of the
action. This means that the robot does not have any location while
moving.