rcll: also require the MPS side as an action resource
We sometimes use a machine side (e.g., C-CS-INPUT) as a resource in the
goal. However, the action parameters always treat the machine and its
side separately. To obtain the resource, check if we can construct a
matching mps-side resource name from two adjacent action parameters and
if so, require that resource as an action resource.