arduino: fix bug when receiving no data from gripper
If waiting for a message is taking too long, the gripper is synchronizing
again, but not calibrating again. This causes the gripper's current position
to be the new origin and is often followed by the gripper trying to overshoot
its boundaries.
The bug is fixed by setting calibrated to false in the handle_nodata function.
This is dangerous since calibration can then start while hovering over an
workpiece and breaking fingers. Therefore, calibration is now starting by
finding the x axis's origin first and is not moving in y and z direction
meanwhile. Now, it is safely executable while hovering over workpieces.