xenonite: require promised resource if formulated based on promise
Import the function `compute-required-resources` from the RCLL domain
and adapt the goal formulation such that we require both the promise
resource and the resource itself if we do *not* formulate based on a
promise. If we formulate based on a promise, only require the promise
resource, which will later be swapped against the actual resource.