xenonite: move waiting for machine processing into start-machine goal
Previously the start-machine goal triggered the creation of a timer
that simulates the extrinsic effect of a machine being done with
operation after a set amount of time. Move this behavior into an
action that is used in the fixed plans of the start-machine goals
to make the effects of promises and resource handovers more prominent.