rcll-central: restrict order production tree formulations actively
Previously, the production trees for all orders that were incoming
where formulated with the goal reasoning component being responsible
for choosing one of the trees. If it could not progress with another
one. With the following addition a transition is started where the
goal reasoning component will be just tasked with selecting the most
suitable goals. To still limit the number of orders produced at the
same time, the formulation of order production trees will be limited.
In a first step, this is just arbitrary selection with a limit of two
active order production trees. Later resource usage, timing, and
order preferences will be considered and the hard limit on the number
of trees will be lifted.