rcll-central: add filter lists that maintain certain selection criteria
Each filter list models a certain set of criteria that a goal has to
fulfill in order to be considered a preferred goal. These criteria are
then individually applied to the set of possible goals -- the ones that
fulfill the base criteria -- and the intersection of the filters is
considered the set of preferred goals. From the preferred goals,
formulate the one with the greatest complexity.