rcll-central: compute workload RS workload for all orders
The previous implementation only asserted workload facts for orders
that made use of the machines. This is insufficient if all orders
should be compared based on their expected contribution to the overall
workload, as orders that are not using a machine don't have the
corresponding facts. This change makes sure that the workload facts
exist for all orders. For orders without interactions, the value is set
to 0.