rcll-central: model the decision for selected orders explicitly
We have two potential choices for orders to pursue. Either there is at
least one that meets all the filters, it is the one with the highest
complexity among those, and it is possible. Or we currently pursue
no order and none meets all the filters, then we pick the one that has
the highest complexity among the possible ones. Introduce two facts,
one storing the possible one with the highest complexity, one storing
the one that meets all filters.

Base the choice for order production tree formulation on these facts
directly.