rcll-central: attempt to fix deadlock at base station on faulty robot
This fix does not work! The problem that was tackled is that if a robot
waits at one side of the BS for another robot to pick up a product on
the other BS side, but that robot never can do it, then the waiting
robot should eventually decide to do something else (or at least get a
chance to do that). The idea was to mark those goals as not being
executable, however it does not fix the issue as the waiting goal does
never terminate (the wait-for-wp simply waits indefinitely, which is a
problem)