cfg: add laptop1 bbsync connections
The laptop on robot 1 is a temporary solution as the computer of robot 1
crashes when using opencv.
The workaround utilizes a laptop that runs the critical plugins of
fawkes. As those require the navgraph it is easiest to simply generate
the navgraph like we do on the bases, via the central agent.