rcll-central: make mount-ring executable for resolvable three-way lock
Add a condition that allows a mount-ring to be executable when the
target-RS output is blocked, if the situation can be resolved through
this mount. I.e., when the current RS' input is blocked and the
WP in the output of the target RS needs to be brought to the current
RS' input.