rcll-central: clean up when failing actions
If an action execution fails normally, the skill and skill-exec-info are
deleted and the action is set to EXECUTION-FAILED.
If our exec monitoring fails an action, the result was different as the
skill-exec-info and skill facts were not cleaned up.

The rules to retry actions were expecting the facts to be present, hence
they did not work if a skill failed normally during execution.