rcll-central: fix condition for order tree creation when no active order
When there is no active order, but there is a possible order that has
already expired, start the expired order in attempts to get more points
(this usually only happens at the end of a game). The previous
condition modelling this, failed to consider MOVE-OUT-OF-WAY goals
which would be active at that point.