rcll-central: create requests for orders
Requests model the need of an order for certain support goals that
are needed to prepare the machines for the assembly goals, which
actually produce the ordered workpiece. These goals include payment
goals for the machines, buffer cap goals to prepare the cap station,
and discards goals, which clean the cap station of the buffer.

By explicitly modelling this demand, the goals needed to achieve them
can be created and managed dynamically based on the current world
state, bookkeeping is enabled, and resolving of situations with
failed goals is simplified.