navgraph-generator-mps: remove wait prefix
The prefix was originally added in order to use the goto feature that
finishes early. However, the goto skill is also capable to drive to
zones directly.
The navgraph point is still useful in order to highlight those zones
which are not near machines. The goto skill on the other hand will work
regardless if the navgraph nodes are present or not.