navgraph-generator-mps: remove wait prefix
The prefix was originally added in order to use the goto feature that
finishes early.
The generation of wait points may differ between the robots and the
central agent, hence a navgraph point may be missing on a robot.
However, the actual target postion shoud be reached anyways.
The goto skill is also capable to drive to zones directly, e.g., by
providing the zone name.
By removing the "WAIT-" prefix the skill will work independent from the
presence of a navgraph point.
The tradeoff is that the current implementation of goto does not finish
early anymore, which was a desired feature. This can be adressed in the
future.