arduino: fix closing gripper
Fixes the bug where the gripper is closing before reaching its destination on
the RC2022.
The bug was caused by a finaling gripper command message while the next command
is still executing which is propably happening now because of higher networking
traffic or damaged cables.

The send_status messages from the arduino firmware were forming a queue
caused by a lower bandwidth between arduino board and robotino base.
The manipulate_wp skill sends in the MOVE_BASE_AND_GRIPPER state a target
gripper position to the gripper, which does not need to be reached, before
starting to only use the gripper_commands in the FINE_TUNE_GRIPPER state.
In this state, the gripper starts moving down while the finish confirmation
from the earlier pose arrives (caused by the delay from the message queue)
and assumes the gripper to be finished with its down movement and
transitions into the next state where the gripper is closing. The gripper is
still continuing to move down and therefore, the gripper crashes with closed
fingers against the workpiece and may break fingers.

To "fix" this issue, we increased the time between each send_status by a factor
of 3 and reduce the traffic between arduino board and robotino base to a degree
where queues are no issue.

A proper solution would require to check if the messages from send_status
belong to the current gripper_command. This could be done by sending the
current target position in the send_status message or by using a command id
which is send by the gripper_command skill and by the send_status messages.